requirements.txt
amqp==5.0.6
appdirs==1.4.4
asgiref==3.4.1
astroid==2.9.0
attrs==21.2.0
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.9
coverage==6.2
Django==3.2.9
django-extensions==3.1.5
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
flake8==3.9.2
fluent-logger==0.10.0
freezegun==1.1.0
gunicorn==20.1.0
isort==5.10.1
itypes==1.2.0
mccabe==0.6.1
mock==4.0.3
msgpack==1.0.3
packaging==21.3
pluggy==1.0.0
psutil==5.8.0
psycopg2-binary==2.9.2
pycodestyle==2.7.0
pyflakes==2.3.0
PyJWT==2.3.0
pylint==2.12.2
pylint-django==2.4.4
pylint-plugin-utils==0.6
pyparsing==3.0.6
pytest==6.2.5
pytest-black==0.3.12
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-flake8==1.0.7
pytest-freezegun==0.4.2
pytest-isort==2.0.0
python-dateutil==2.8.2
pytz==2021.3
regex==2021.11.10
requests==2.26.0
sentry-sdk==1.5.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3